made the photocontrol server app fully 64bit compatible removed the sscrypto framework, because encryption is not (yet?) needed for this app

+++ classes/DirectoryHandler.m
@@ -45,7 +45,7 @@
 			if(![component isEqual:@"/"])
 			{
 				NSLog(@"%@", component);
-				int intIndex;
+				NSInteger intIndex;
 				BOOL success = [[NSScanner scannerWithString:component] scanInteger:&intIndex];
 				if(!success)
 				{
